Backend/requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
django-summernote==0.8.11.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.14.0
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.0
idna==2.10
inflection==0.5.1
jsonschema==3.2.0
jwcrypto==0.8
oauth2client==4.1.3
Pillow==8.1.2
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.1
pyparsing==2.4.7
Pyrebase4==4.4.3
pyrsistent==0.17.3
python-decouple==3.4
python-jwt==3.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0

